+++ jest.setupAfterEnv.js
@@ -5,14 +5,6 @@ const parsedCdsOptions = JSON.parse(process.env.CDS_CONFIG ?? "{}");
 parsedCdsOptions.requires ??= {};
 parsedCdsOptions.requires.outbox = "persistent-outbox";
 
-if (/true/i.test(process.env.OLD_DB_SERVICE)) {
-  parsedCdsOptions.requires.db = {
-    kind: "legacy-sqlite",
-    credentials: {
-      url: ":memory:",
-    },
-  };
-}
 process.env.CDS_CONFIG = JSON.stringify(parsedCdsOptions);
 
 // turn off regular and error logging;